#1 Add python support

+++ Tests/Utils.hpp
@@ -6,12 +6,12 @@
 
 struct FileGuard
 {
-	FileGuard(std::filesystem::path path) : m_Path(std::move(path)) {}
+	FileGuard(std::filesystem::path path, std::string params = {}) : m_Path(std::move(path)), m_Params(std::move(params)) {}
 	~FileGuard()
 	{
 		try
 		{
-			if (!m_Path.empty() && std::filesystem::exists(m_Path))
+			if (m_bDelete && !m_Path.empty() && std::filesystem::exists(m_Path))
 				std::filesystem::remove(m_Path);
 		}
 		catch (const std::exception& ex)
@@ -20,13 +20,17 @@ struct FileGuard
 		}
 	}
 
-	const std::filesystem::path& Path() const { return m_Path; }
-	std::string Extension() const { return m_Path.extension().string().erase(0, 1); }
+	void PreventDeletion() { m_bDelete = false; }
 
+	const std::filesystem::path& Path() const { return m_Path; }
+	const std::string& Params() const { return m_Params; }
+	std::string Extension() const { return m_Path.extension().string().erase(0, 1); }
 	operator std::filesystem::path() const { return m_Path; }
 
 private:
 	std::filesystem::path m_Path;
+	std::string m_Params;
+	bool m_bDelete = true;
 };
 
 inline FileGuard CompileCode(std::string code, const FileGuard& path)
@@ -34,5 +38,5 @@ inline FileGuard CompileCode(std::string code, const FileGuard& path)
 	Compiler compiler(std::move(code), path, Compiler::MakeImplFromExtension(path.Extension()));
 	EXPECT_TRUE(compiler.Run());
 	EXPECT_EQ(compiler.Error().size(), 0) << "Compilation fails with the following message: " << compiler.Error();
-	return { compiler.ExecutablePath() };
+	return { compiler.ExecutablePath(), compiler.ExecutableParams() };
 }
